fix(diff): move dev text helper out of renderer tree
Keep Node-only dev diff text reads under config and update the Vite import.

Also validates rename sources and corrects the synthesized diff comment from review.
