+++ tsconfig.node.json
@@ -18,5 +18,5 @@
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true
   },
-  "include": ["vite.config.ts", "vite-plugin-files.ts"]
+  "include": ["vite.config.ts", "vite-plugin-files.ts", "config/**/*.ts"]
 }
